+++ R/raster-info.R
@@ -24,6 +24,13 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
   datavars$subdataset[sds]
 }
 
+
+.gt_dim_to_extent <- function (x, dim)
+{
+  xx <- c(x[1], x[1] + dim[1] * x[2])
+  yy <- c(x[4] + dim[2] * x[6], x[4])
+  c(xx, yy)
+}
 #' Raster information
 #'
 #' Return the basic structural metadata of a raster source understood by GDAL.
@@ -33,6 +40,7 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
 #' The structural metadata are
 #'
 #' \describe{
+#' \item{extent}{the extent of the data, xmin, xmax, ymin, ymax - these are the lower left and upper right corners of pixels}
 #' \item{geotransform}{the affine transform}
 #' \item{dimXY}{dimensions x-y, columns*rows}
 #' \item{minmax}{numeric values of the computed min and max from the first band (optional)}
@@ -43,7 +51,12 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
 #' \item{nodata_value}{not implemented}
 #' \item{overviews}{the number and size of any available overviews}
 #' \item{filelist}{the list of files involved (may be none, and so will be a single NA character value)}
+#'
 #' }
+#'
+#' Note that the geotransform is a kind of obscure combination of the extent and dimension, I don't find it
+#' useful and modern GDAL is moving away from needing it so much. Extent is more sensible and used in many places in
+#' a straightforward way.
 #'
 #' On access vapour functions will report on the existence of subdatasets while
 #' defaulting to the first subdataset found.
@@ -127,7 +140,9 @@ sds_boilerplate_checks <- function(x, sds = NULL) {
 #' vapour_raster_info(f)
 vapour_raster_info <- function(x, ..., sds = NULL, min_max = FALSE) {
   datasourcename <- sds_boilerplate_checks(x, sds = sds)
-  raster_info_gdal_cpp(dsn = datasourcename, min_max = min_max)
+  info <- raster_info_gdal_cpp(dsn = datasourcename, min_max = min_max)
+  info[["extent"]] <- .gt_dim_to_extent(info$geotransform, info$dimXY)
+  info
 }
 
 #' Raster ground control points