+++ NEWS.md
@@ -1,6 +1,7 @@
 # vapour 0.7.5
 
-* Fixed input type problems found on CRAN, see https://github.com/hypertidy/vapour/issues/110
+* Fixed input type problems found on CRAN, see https://github.com/hypertidy/vapour/issues/110. Thanks to CRAN 
+ for explaining. 
 
 * Warper input can now include multiple sources.